// deploy/restpe-contabilidad-front-end/src/app/modules/rrhh/domain/models/categoria-laboral.entity.ts
export interface CategoriaLaboralEntity {
  categoria_laboral_codigo: string;
  categoria_laboral_nombre: string;
  categoria_laboral_tipo: string;
  categoria_laboral_beneficio: string;
  categoria_laboral_retenciones: string;
  categoria_laboral_aportes: string;
  categoria_laboral_fecha_creacion: string;
  categoria_laboral_vigencia: string;
  categoria_laboral_estado: string;
  categoria_laboral_trabajadores?: string[];
}