// deploy/restpe-contabilidad-front-end/src/app/modules/rrhh/domain/models/datos-personales.entity.ts
export interface DatosPersonalesEntity {
  empleado_codigo: string;
  empleado_nombres_apellidos: string;
  empleado_tipo_documento: string;
  empleado_documento: string;
  empleado_fecha_nacimiento: string;
  empleado_sexo: string;
  empleado_nacionalidad: string;
  empleado_pais: string;
  empleado_departamento: string;
  empleado_provincia: string;
  empleado_distrito: string;
  empleado_direccion: string;
  empleado_telefono_privado: string;
  empleado_correo_privado: string;
  empleado_contacto_emergencia_nombre: string;
  empleado_contacto_emergencia_telefono: string;
  empleado_fecha_inicio: string;
  empleado_fecha_fin: string;
  empleado_tipo_contrato: string;
  empleado_cargo: string;
  empleado_sucursal: string;
  empleado_jornada: string;
  empleado_modalidad: string;
  empleado_centro_costos: string;
  empleado_regimen_salud: string;
  empleado_regimen_pensionario: string;
  empleado_moneda: string;
  empleado_remuneracion: string;
  liquidacion_asignacion_familiar?: number;
  liquidacion_promedio_gratificacion?: number;
  empleado_periodicidad_pago: string;
  empleado_estado: string;
  empleado_ingresos?: number;
  empleado_descuentos?: number;
  empleado_aportes?: number;
  empleado_neto?: number;
  aporteempleado?: number;
  diasLaborados?: number;
  diasSubsidiados?: number;
  diasNoLaborados?: number;
  diasTotal?: number;
  ordinaria?: string;
  ordinaria2?: string;
  ordinariaH?: string;
  ordinariaM?: string;
  sobretiempo?: string;
  sobretiempo2?: string;
  sobretiempoH?: string;
  sobretiempoM?: string;
  totalHoras?: string;
  totalHoras2?: string;
  totalHorasH?: string;
  totalHorasM?: string;
  totalMinutos?: string;
}